+++ data-transfer/src/main/java/it/eng/datatransfer/model/AbstractTransferMessage.java
@@ -1,5 +1,6 @@
 package it.eng.datatransfer.model;
 
+import java.io.Serializable;
 import java.util.UUID;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -8,8 +9,10 @@ import com.fasterxml.jackson.annotation.JsonProperty.Access;
 import it.eng.tools.model.DSpaceConstants;
 import jakarta.validation.constraints.NotNull;
 
-public abstract class AbstractTransferMessage {
+public abstract class AbstractTransferMessage implements Serializable {
 	
+	private static final long serialVersionUID = -3150306747585657302L;
+
 	@JsonProperty(value = DSpaceConstants.CONTEXT, access = Access.READ_ONLY)
 	private String context = DSpaceConstants.DATASPACE_CONTEXT_0_8_VALUE;
 
@@ -27,4 +30,8 @@ public abstract class AbstractTransferMessage {
 	protected String createNewId() {
 	    return UUID.randomUUID().toString();
 	}
+	
+    protected String createNewPid() {
+        return "urn:uuid:" + UUID.randomUUID();
+    }
 }