+++ data-transfer/src/main/java/it/eng/datatransfer/repository/TransferProcessRepository.java
@@ -1,5 +1,6 @@
 package it.eng.datatransfer.repository;
 
+import java.util.Collection;
 import java.util.Optional;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
@@ -13,4 +14,8 @@ public interface TransferProcessRepository extends MongoRepository<TransferProce
     Optional<TransferProcess> findByConsumerPidAndProviderPid(String consumerPid, String providerPid);
     
     Optional<TransferProcess> findByProviderPid(String providerPid);
+    
+    Optional<TransferProcess> findByAgreementId(String agreementId);
+    
+    Collection<TransferProcess> findByState(String state);
 }