+++ data-transfer/src/test/java/it/eng/datatransfer/rest/protocol/ProviderDataTransferControllerTest.java
@@ -2,6 +2,8 @@ package it.eng.datatransfer.rest.protocol;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.isNull;
 import static org.mockito.Mockito.when;
 
 import org.junit.jupiter.api.DisplayName;
@@ -11,9 +13,15 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 
-import it.eng.datatransfer.exceptions.TransferProcessNotFound;
+import com.fasterxml.jackson.databind.JsonNode;
+
+import it.eng.datatransfer.exceptions.TransferProcessExistsException;
+import it.eng.datatransfer.exceptions.TransferProcessNotFoundException;
 import it.eng.datatransfer.model.Serializer;
+import it.eng.datatransfer.model.TransferRequestMessage;
+import it.eng.datatransfer.model.TransferStartMessage;
 import it.eng.datatransfer.service.DataTransferService;
 import it.eng.datatransfer.util.MockObjectUtil;
 import jakarta.validation.ValidationException;
@@ -39,16 +47,18 @@ public class ProviderDataTransferControllerTest {
 	@DisplayName("Get TransferProcess for ProviderPid - not found")
 	public void transferProcessNtFound() {
 		when(dataTransferService.findTransferProcessByProviderPid(MockObjectUtil.PROVIDER_PID))
-			.thenThrow(new TransferProcessNotFound("Not found"));
-		assertThrows(TransferProcessNotFound.class, 
+			.thenThrow(new TransferProcessNotFoundException("Not found"));
+		assertThrows(TransferProcessNotFoundException.class, 
 				()-> controller.getTransferProcessByProviderPid(MockObjectUtil.PROVIDER_PID).getStatusCode());
 	}
 	
 	@Test
 	@DisplayName("Initiate TransferProcess")
 	public void initateDataTransfer() {
-		assertEquals(HttpStatus.NOT_IMPLEMENTED, 
-				controller.initiateDataTransfer(Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_REQUEST_MESSAGE)).getStatusCode());
+		when(dataTransferService.initiateDataTransfer(any(TransferRequestMessage.class)))
+			.thenReturn(MockObjectUtil.TRANSFER_PROCESS_REQUESTED);
+		ResponseEntity<JsonNode> response = controller.initiateDataTransfer(Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_REQUEST_MESSAGE));
+		assertEquals(HttpStatus.CREATED, response.getStatusCode());
 	}
 	
 	@Test
@@ -60,19 +70,43 @@ public class ProviderDataTransferControllerTest {
 	}
 	
 	@Test
+	@DisplayName("Initiate TransferProcess - service error")
+	public void initateDataTransfer_service_error() {
+		when(dataTransferService.initiateDataTransfer(any(TransferRequestMessage.class)))
+			.thenThrow(new TransferProcessExistsException("message", MockObjectUtil.CONSUMER_PID));
+		assertThrows(TransferProcessExistsException.class, () ->
+			controller.initiateDataTransfer(Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_REQUEST_MESSAGE))
+		);
+	}
+	
+	@Test
 	@DisplayName("Start TransferProcess")
 	public void startDataTransfer() {
-		assertEquals(HttpStatus.NOT_IMPLEMENTED, 
-				controller.startDataTransfer(MockObjectUtil.PROVIDER_PID,
-						Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_START_MESSAGE)).getStatusCode());
+	when(dataTransferService.startDataTransfer(any(TransferStartMessage.class), isNull(), any(String.class)))
+		.thenReturn(MockObjectUtil.TRANSFER_PROCESS_STARTED);
+		
+	ResponseEntity<JsonNode> response = controller.startDataTransfer(MockObjectUtil.PROVIDER_PID,
+			Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_START_MESSAGE));
+	
+	assertEquals(HttpStatus.OK, response.getStatusCode());
 	}
 	
 	@Test
 	@DisplayName("Start TransferProcess - invalid request body")
 	public void startDataTransfer_invalidBody() {
 		assertThrows(ValidationException.class, () ->
-			controller.startDataTransfer(MockObjectUtil.PROVIDER_PID, Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_COMPLETION_MESSAGE))
-		);
+			controller.startDataTransfer(MockObjectUtil.PROVIDER_PID, 
+					Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_COMPLETION_MESSAGE)));
+	}
+	
+	@Test
+	@DisplayName("Start TransferProcess - error service")
+	public void startDataTransfer_errorService() {
+		when(dataTransferService.startDataTransfer(any(TransferStartMessage.class), isNull(), any(String.class)))
+			.thenThrow(new TransferProcessNotFoundException("TransferProcess not found test"));
+		assertThrows(TransferProcessNotFoundException.class, () ->
+			controller.startDataTransfer(MockObjectUtil.PROVIDER_PID, 
+					Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_START_MESSAGE)));
 	}
 	
 	@Test