+++ tools/src/main/java/it/eng/tools/model/IConstants.java
@@ -1,0 +1,15 @@
+package it.eng.tools.model;
+
+public interface IConstants {
+
+	public static final String ID = "_id";
+	public static final String INSERT_DATE = "insertDate";
+	public static final String KEY = "key";
+	public static final String LAST_CHANGE_DATE = "lastChangeDate";
+	public static final String MANDATORY = "mandatory";
+	public static final String SAMPLE_VALUE = "sampleValue";
+	public static final String TYPE = "type";
+	public static final String VALUE = "value";
+	
+	
+}